LayoutTests/platform/chromium/inspector/timeline/timeline-animation-frame-expected.txt

Tests the Timeline events for Animation Frame feature

RegisterAnimationFrameCallback Properties:
{
    startTime : <number>
    stackTrace : <object>
    data : {
        id : <number>
    }
    type : "RegisterAnimationFrameCallback"
    usedHeapSize : <number>
    totalHeapSize : <number>
}
FireAnimationFrameEvent Properties:
{
    startTime : <number>
    stackTrace : <object>
    data : {
        id : <number>
    }
    children : <object>
    endTime : <number>
    type : "FireAnimationFrameEvent"
    usedHeapSize : <number>
    totalHeapSize : <number>
}
CancelAnimationFrameCallback Properties:
{
    startTime : <number>
    stackTrace : <object>
    data : {
        id : <number>
    }
    type : "CancelAnimationFrameCallback"
    usedHeapSize : <number>
    totalHeapSize : <number>
}
